backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-docx==1.1.2
langchain==0.3.13
langchain-google-genai==2.0.7
langchain-chroma==0.1.4
langchain-groq==0.2.2
langchain-community==0.3.13
python-dotenv==1.0.1
